Remove .SHELLFLAGS variable
With make 4.0 and 4.1 the .SHELLFLAGS variable changes the behavior
of make and causes a "No such file or directory" error:

   wink@desktop:~/prgs/chisel/chisel-tutorial-fork/examples$ make Parity.out
   set -e -o pipefail; sbt -Dsbt.log.noformat=true "run Parity --genHarness --compile --test --backend c " | tee Parity.out
   /bin/sh: set -e -o pipefail; sbt -Dsbt.log.noformat=true "run Parity --genHarness --compile --test --backend c " | tee Parity.out: No such file or directory
   ../suffix.mk:66: recipe for target 'Parity.out' failed
   make: *** [Parity.out] Error 1


The I simplified the "%.out" target so that it was just an empty $(SBT):

  %.out: %.scala
           $(SBT)

And then the true reason for the "No such file or directory" is that
sbt-launcher-lib.bash script can not be found because the current working
directory is "incorrect":

   wink@desktop:~/prgs/chisel/chisel-tutorial-fork/examples$ make Parity.out
   sbt
   /opt/sbt/bin/sbt: line 59: /home/wink/prgs/chisel/chisel-tutorial-fork/examples/sbt-launch-lib.bash: No such file or directory
   ../suffix.mk:64: recipe for target 'Parity.out' failed
   make: *** [Parity.out] Error 1


To me this seems to be a bug with make but we're not going to change make
so the best solution seems to be just remove the .SHELLFLAGS variable.
With it removed the code works for both make 3.81, 4.0 and 4.1.